+++ backend/src/main/java/com/team5/issue_tracker/comment/dto/CommentResponse.java
@@ -1,4 +1,4 @@
-package com.team5.issue_tracker.common.comment.dto;
+package com.team5.issue_tracker.comment.dto;
 
 import java.time.Instant;
