+++ backend/src/main/java/com/team5/issue_tracker/comment/service/CommentService.java
@@ -1,11 +1,11 @@
-package com.team5.issue_tracker.common.comment.service;
+package com.team5.issue_tracker.comment.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.team5.issue_tracker.common.comment.domain.Comment;
-import com.team5.issue_tracker.common.comment.dto.CommentRequest;
-import com.team5.issue_tracker.common.comment.reqository.CommentRepository;
+import com.team5.issue_tracker.comment.domain.Comment;
+import com.team5.issue_tracker.comment.dto.CommentRequest;
+import com.team5.issue_tracker.comment.reqository.CommentRepository;
 
 import lombok.RequiredArgsConstructor;
