+++ backend/src/main/java/com/team5/issue_tracker/issue/dto/response/IssueDetailResponse.java
@@ -3,7 +3,7 @@ package com.team5.issue_tracker.issue.dto.response;
 import java.time.Instant;
 import java.util.List;
 
-import com.team5.issue_tracker.common.comment.dto.CommentResponse;
+import com.team5.issue_tracker.comment.dto.CommentResponse;
 import com.team5.issue_tracker.label.dto.response.LabelResponse;
 import com.team5.issue_tracker.milestone.dto.response.MilestoneResponse;
 import com.team5.issue_tracker.user.dto.UserSummaryResponse;
