+++ backend/src/main/java/com/team5/issue_tracker/issue/mapper/IssueMapper.java
@@ -3,7 +3,7 @@ package com.team5.issue_tracker.issue.mapper;
 import java.util.List;
 import java.util.Map;
 
-import com.team5.issue_tracker.common.comment.dto.CommentResponse;
+import com.team5.issue_tracker.comment.dto.CommentResponse;
 import com.team5.issue_tracker.issue.dto.IssueQueryDto;
 import com.team5.issue_tracker.issue.dto.response.IssueBaseResponse;
 import com.team5.issue_tracker.issue.dto.response.IssueDetailResponse;
