+++ backend/src/main/java/com/team5/issue_tracker/issue/query/IssueQueryService.java
@@ -5,12 +5,12 @@ import java.util.Map;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import com.team5.issue_tracker.common.comment.query.CommentQueryRepository;
+import com.team5.issue_tracker.comment.query.CommentQueryRepository;
 import com.team5.issue_tracker.issue.dto.IssueQueryDto;
 import com.team5.issue_tracker.issue.dto.IssueSearchCondition;
 import com.team5.issue_tracker.issue.dto.request.IssueSearchRequest;
 import com.team5.issue_tracker.issue.dto.response.IssueBaseResponse;
-import com.team5.issue_tracker.common.comment.dto.CommentResponse;
+import com.team5.issue_tracker.comment.dto.CommentResponse;
 import com.team5.issue_tracker.issue.dto.response.IssueCountResponse;
 import com.team5.issue_tracker.issue.dto.response.IssueDetailResponse;
 import com.team5.issue_tracker.issue.dto.response.IssuePageResponse;
